fix workaround of zsync bug

+++ useful-tools/uruntime2appimage.sh
@@ -146,7 +146,7 @@ if [ -n "$UPINFO" ]; then
 	zsyncmake -u "$OUTNAME" "$OUTPATH"/"$OUTNAME"
 
 	# there is a nasty bug that zsync make places the .zsync file in PWD
-	if [ -f "$OUTNAME".zsync ]; then
+	if [ ! -f "$OUTPATH"/"$OUTNAME".zsync ] && [ -f "$OUTNAME".zsync ]; then
 		mv "$OUTNAME".zsync "$OUTPATH"/"$OUTNAME".zsync
 	fi
 fi
